Modify : sledAudioEffect -> sledFrictionAudioEffect
함수 생성 및 실행 위치 변경.
sound는 soundmanager로
